tring to pass last stage

+++ src/indexParser.cpp
@@ -217,7 +217,7 @@ void printIndexLeafPage(std::vector<char> &database_file , unsigned short rootPa
 
 
 
-    for(int i = 0 ; i < database_file.size();i+=4096)
+    for(unsigned long i = 0 ; i < database_file.size();i+=4096)
     {
     unsigned long pageStart = i;
